requirements/test.txt
case>=1.3.1
pytest==4.0.2
pytest-django==3.4.4
pytz>dev
pytest-cov==2.6.1
codecov==2.0.11
